requirements.txt
asttokens==2.4.1
backcall==0.2.0
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
coloredlogs==15.0.1
cycler==0.12.1
decorator==5.1.1
executing==2.1.0
Flask==2.2.3
flatbuffers==24.3.25
fonttools==4.54.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
ipython==8.4.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.5.2
matplotlib-inline==0.1.7
mpmath==1.3.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
onnx==1.16.1
onnxruntime==1.14.1
onnxruntime-gpu==1.19.2
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.64
packaging @ file:///croot/packaging_1720101850331/work
pandas==1.1.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
ply==3.11
prompt_toolkit==3.0.48
protobuf==5.28.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.15
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
pyserial==3.5
PySide2==5.15.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0
requests==2.27.1
scipy==1.13.1
seaborn==0.11.2
shiboken2==5.15.2.1
sip @ file:///croot/sip_1698675935381/work
six==1.16.0
stack-data==0.6.3
sympy==1.13.3
tensorrt @ file:///opt/TensorRT-8.6.1.6/python/tensorrt-8.6.1-cp39-none-linux_x86_64.whl#sha256=872dab169be90c72d1edee297670ad9127bea55c3360e8edce33f50a4668f826
tensorrt-cu12==10.6.0
tensorrt-cu12-bindings==10.6.0
tensorrt-cu12-libs==10.6.0
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.0
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
